train/requirements.txt
numpy==1.14.0
scipy==0.19.1
torch==1.1.0
tensorflow==1.13.1
vispy==0.5.3
torchvision==0.2.2.post3
opencv_contrib_python==4.1.0.25
matplotlib==2.2.3
Pillow==6.1.0
PyYAML==5.1.1